Write DHCP offer to static Netplan file:
This is useful when the host being provisioned
will run the Boots DHCP server. This happens
when using EKS Anywhere which colocates Boots with
all CAPI Tinkerbell components. These components
are initially created on a Bootstrap cluster, then
a permanent management cluster is created, and all
components, including Boots, are pivoted to this
permanent management cluster.

Signed-off-by: Jacob Weinstock <jakobweinstock@gmail.com>